+++ locales/es-ar/passwordRecovery.json
@@ -3,5 +3,6 @@
     "userIdentifier": "Usuario o email",
     "instructions": "Escribí el usuario o el correo electrónico, y te enviaremos a tu casilla las instrucciones para reestablecer tu contraseña. Cualquier problema podés comunicarte a",
     "mailSent": "Si tu usuario tiene un correo registrado, recibirás en tu casilla las instrucciones. Recordá revisar tu carpeta de spam. Cualquier problema podés comunicarte a",
-    "serverError": "Ocurrió un error, vuelva a intentar"
+    "serverError": "Ocurrió un error, vuelva a intentar",
+    "backToHome": "Regresar al inicio"
 }
